Hide tmlog_appender class
