+++ src/noir/common/log.h
@@ -1,20 +1,7 @@
 #pragma once
-#include <fc/log/appender.hpp>
-#include <fc/log/log_message.hpp>
 #include <fc/log/logger_config.hpp>
 
 namespace noir::log {
-
-class tmlog_appender : public fc::appender {
-public:
-  explicit tmlog_appender(const fc::variant& args);
-  tmlog_appender();
-
-  virtual ~tmlog_appender();
-
-  virtual void initialize(boost::asio::io_service& io_service) override;
-  virtual void log(const fc::log_message& m) override;
-};
 
 static const char* DEFAULT_LOGGER_NAME = "TMLOG";
 void initialize(const char* logger_name);
